포인트 환전 및 성공시 마이페이지 리로드

+++ app/src/main/java/com/ajou/capstone_design_freitag/UI/dto/User.java
@@ -25,7 +25,7 @@ public class User {
     private int visit;
     private String totalPoint;
     private int point;
-    private int accuracy;
+    private double accuracy;
     private String level;
     private int bank;
 
@@ -116,11 +116,11 @@ public class User {
         this.point = point;
     }
 
-    public int getAccuracy() {
+    public double getAccuracy() {
         return accuracy;
     }
 
-    public void setAccuracy(int accuracy) {
+    public void setAccuracy(double accuracy) {
         this.accuracy = accuracy;
     }
